:recycle: refactor: 도메인 모델에 관한 예외는 도메인 패키지로 분류함
